[feat] Sheet Binding with Views #44 (#65)
